feat(removed search from nav, created tabs comp): removed Search From Nav, Created Tabs Comp
Removed Search From Nav, Created Tabs Comp
