+++ .vs/VSWorkspaceState.json
@@ -1,0 +1,8 @@
+{
+  "ExpandedNodes": [
+    "",
+    "\\server"
+  ],
+  "SelectedNode": "\\client",
+  "PreviewInSolutionExplorer": false
+}
