+++ client/trademon/src/components/aaPageComponents/dmPage.tsx
@@ -1,5 +1,5 @@
 import React from 'react'
-import SearchBar from '../navComponents/searchComponents/searchBarComponent'
+import NavComponent from '../navComponents/navComponent'
 import DMSummaryTile from '../tileComponents/myProfileTileComponents/dmSummaryTile'
 import '../../styling/containers.scss'
 
@@ -10,7 +10,7 @@ export default function DMPage() {
   return (
     <div className="dm-page">
       <div>
-        <SearchBar></SearchBar>
+        <NavComponent></NavComponent>
       </div>
       <div className="menu-title">
         My Trades
