+++ client/trademon/src/components/aaPageComponents/myProfilePage.tsx
@@ -2,7 +2,7 @@ import React from 'react'
 import ProfileBanner from '../tileComponents/myProfileTileComponents/profileBannerComponent'
 import WatchList from '../tileComponents/myProfileTileComponents/watchListComponent'
 import MyOffers from '../tileComponents/myProfileTileComponents/myOffersComponent'
-import SearchBar from '../navComponents/searchComponents/searchBarComponent'
+import NavComponent from '../navComponents/navComponent'
 import Footer from '../navComponents/footer';
 import '../../styling/containers.scss'
 
@@ -11,7 +11,7 @@ export default function MyProfile() {
 
   return (
     <div>
-      <SearchBar></SearchBar>
+      <NavComponent></NavComponent>
       <div className="menu-title">
         My Profile
       </div>
