--- client/trademon/src/components/navComponents/mobileNavComponents/placeholder.tsx
@@ -1,8 +1,0 @@
-import React from 'react'
-import '../../styling/forms.scss'
-export default function Placeholder() {
-  return (
-    <div>
-    </div>
-  )
-}
