+++ client/trademon/src/styling/containers.scss
@@ -86,7 +86,7 @@
   margin-top: 5vh;
   display: flex;
   flex-wrap: wrap;
-  width: 95%;
+  width: 100vw;
   justify-content: space-evenly;
 }
