+++ client/trademon/src/styling/demo.scss
@@ -10,8 +10,8 @@
     bottom: 0;
     z-index: 5;
     background-color: black;
-    width: 20%;
-    height: 8vh;
+    width: 18%;
+    height: 10vh;
   }
   .demo-content {
     display: flex;
